# [v6r20] FileManager._getRepIDsForReplica: faster query