@zmathe 2018-06-01
@andresailer The indexing of the table is not optimal. The original query must be faster, if you have a composite index on (FileID, and replicaid) and must have another composite index on FileID and Storage id (SEID). Could you please try to add an index and check the execution plan?