@andresailer 2018-06-01
There is an index on FileID and SEID, but it is not used.
```
mysql> show create table FC_Replicas;
+-------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Table       | Create Table                                                                                                                                                                                                                                                                                                                                                                      |
+-------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| FC_Replicas | CREATE TABLE `FC_Replicas` (
  `RepID` int(11) NOT NULL AUTO_INCREMENT,
  `FileID` int(11) NOT NULL,
  `SEID` int(11) NOT NULL,
  `Status` smallint(5) unsigned NOT NULL,
  PRIMARY KEY (`RepID`),
  KEY `FileID` (`FileID`),
  KEY `SEID` (`SEID`),
  KEY `FileID_2` (`FileID`,`SEID`),
  KEY `Status` (`Status`)
) ENGINE=InnoDB AUTO_INCREMENT=62758131 DEFAULT CHARSET=latin1 |
+-------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
```