@andresailer 2018-06-01
This explanation seems relevant, https://stackoverflow.com/a/9597294 , 
```
If query selects only fields from index (or if table has no other fields) by composite WHERE ... IN, index will be used:

SELECT a,b FROM `table` WHERE (a, b) IN ((1,2), (2,4), (1,3))

Otherwise it will not be used. The workaround is to use derived query:

SELECT t.* FROM (SELECT a, b FROM `table` WHERE (a, b) IN ((1,2), (2,4), (1,3))) AS o INNER JOIN `table` AS t ON (t.a = o.a AND t.b = o.b)
```
```
but even if I just select FileID,SEID the EXPLAIN... wants to go through all rows.
```
mysql> EXPLAIN SELECT FileID,SEID FROM FC_Replicas  WHERE (FileID,SEID) IN ((38555741,6),(38555931,6),(38555834,6),(38555833,6),(38555777,6),(38555745,6),(38555775,6),(38555643,6),(38555916,6),(38555637,6));
+----+-------------+-------------+-------+---------------+----------+---------+------+----------+--------------------------+
| id | select_type | table       | type  | possible_keys | key      | key_len | ref  | rows     | Extra                    |
+----+-------------+-------------+-------+---------------+----------+---------+------+----------+--------------------------+
|  1 | SIMPLE      | FC_Replicas | index | NULL          | FileID_2 | 8       | NULL | 11651389 | Using where; Using index |
+----+-------------+-------------+-------+---------------+----------+---------+------+----------+--------------------------+
1 row in set (0.00 sec)
```