@zmathe 2018-06-01
Thanks for the explanation. It is clear now.  It means only the following composite index can help FC_Replicas(FileID,SEID, reqID) on that order... In that case we do not need FileID_2 index, and we have almost everything indexed. But maybe this you do not want.