@fstagni 2018-06-04
Which version of MySQL are you using?