@andresailer 2018-06-04
```
mysql> SELECT @@version;
+------------+
| @@version  |
+------------+
| 5.6.17-log |
+------------+
```
I don't have the FileCatalogDB on the dev instance