@fstagni 2018-06-04
@zmathe do you have a better solution than the one proposed? Or approve?