@andresailer 2018-06-04
I am currently dumping my FC_Replicas table into the dev instance with mysql 5.7 to see if things are working better there. 
It would also be interesting to know what happens on MariaDB?