@andresailer 2018-06-05
On mysql 5.7 things are working properly with this change in the query.
```
mysql> SELECT @@version;
+------------+
| @@version  |
+------------+
| 5.7.15-log |
+------------+
1 row in set (0.00 sec)

mysql> EXPLAIN SELECT RepID,FileID,SEID FROM FC_Replicas WHERE (FileID,SEID)  IN ((38555741,3),(38555931,3),(38555834,3),(38555833,3),(38555777,3),(38555745,2),(38555775,5),(38555643,5),(38555916,5),(38555637,5));
+----+-------------+-------------+------------+-------+----------------------+----------+---------+------+------+----------+--------------------------+
| id | select_type | table       | partitions | type  | possible_keys        | key      | key_len | ref  | rows | filtered | Extra                    |
+----+-------------+-------------+------------+-------+----------------------+----------+---------+------+------+----------+--------------------------+
|  1 | SIMPLE      | FC_Replicas | NULL       | range | FileID,SEID,FileID_2 | FileID_2 | 8       | NULL |   10 |   100.00 | Using where; Using index |
+----+-------------+-------------+------------+-------+----------------------+----------+---------+------+------+----------+--------------------------+
1 row in set, 1 warning (0.00 sec)

mysql> SELECT RepID,FileID,SEID FROM FC_Replicas WHERE (FileID,SEID)  IN ((38555741,3),(38555931,3),(38555834,3),(38555833,3),(38555777,3),(38555745,2),(38555775,5),(38555643,5),(38555916,5),(38555637,5));
Empty set (0.01 sec)

```