@zmathe 2018-06-05
@fstagni : I am not favour of changing the query. Let me explain why: The optimiser in most of new  MySQL release can be changed, which can be good or bad. The query which works one case does not work in other case. It can happen a better filtering will be introduced and because of this change, the advantages of the new feature will not work. Next time we may have another query which need this kind of fix. I always try to fix in the db level, changing the index or creating new index, this does not require change in the code and you always can drop an index.
I would try using the proposed index: FC_Replicas(FileID,SEID, reqID) even you can keep FileID_2 instead of modifying the code. For me the modification of the code always the last.