@zmathe 2018-06-05
@andresailer Could you please create a new index FC_Replicas(FileID,SEID, reqID) and test the query using 5.6 and 5.7?