@andresailer 2018-06-05
Sorry, it takes to long to re-index the table.

If no-one else has this issue, I will keep this as a hot-fix and drop it once I move to mysql 5.7